make types util
