+++ creator-node/src/apiHelpers.ts
@@ -1,6 +1,7 @@
 import type { Request, Response, NextFunction } from 'express'
 import type { AxiosResponse } from 'axios'
 import type Logger from 'bunyan'
+import type { ApiResponse, CustomRequest } from './utils'
 
 import { tracing } from './tracer'
 
@@ -12,28 +13,6 @@ import {
   logger as genericLogger
 } from './logging'
 import { generateTimestampAndSignature } from './apiSigning'
-
-// Content node app adds additional fields to the Express request object. This typing
-// is a type that adds additional fields to the request object.
-export type CustomRequest = Request & {
-  logger: Logger
-  normalizedPath?: string
-}
-
-export type LogContext = {
-  requestID: string
-}
-
-type ApiResponse = {
-  statusCode: number
-  object: {
-    error?: any
-    timestamp?: Date
-    signature?: string
-    data?: any
-    signer?: string
-  }
-}
 
 export const handleResponse = <
   TAsyncFunction extends (...args: any[]) => Promise<any>
