+++ creator-node/src/components/tracks/tracksComponentService.ts
@@ -1,4 +1,4 @@
-import type { LogContext } from '../../apiHelpers'
+import type { LogContext } from '../../utils'
 import {
   logger as genericLogger,
   logInfoWithDuration,
