+++ creator-node/src/utils/types.ts
@@ -1,0 +1,26 @@
+import type Logger from 'bunyan'
+import type { Request } from 'express'
+
+export type ValuesOf<T> = T[keyof T]
+
+// Content node app adds additional fields to the Express request object. This typing
+// is a type that adds additional fields to the request object.
+export type CustomRequest = Request & {
+  logger: Logger
+  normalizedPath?: string
+}
+
+export type LogContext = {
+  requestID: string
+}
+
+export type ApiResponse = {
+  statusCode: number
+  object: {
+    error?: any
+    timestamp?: Date
+    signature?: string
+    data?: any
+    signer?: string
+  }
+}
