+++ creator-node/src/utils/utils.ts
@@ -2,8 +2,6 @@ import { recoverPersonalSignature } from 'eth-sig-util'
 import { logger as genericLogger } from '../logging'
 import config from '../config'
 
-export type ValuesOf<T> = T[keyof T]
-
 export function verifySignature(data: any, sig: any) {
   return recoverPersonalSignature({ data, sig })
 }
